# openapi.yaml
openapi: 3.0.3
...
info:
  title: RTK Account Manager API
  version: 0.1.0
  description: |
    RTK Account Manager is a backend service API and authoritative account-side
    control plane for identity, tenant context, authorization, device registry,
    and provisioning intent.

    This service does not provide a WebUI, dashboard, frontend console, BFF
    session layer, UI preferences, or display-cache ownership. WebUI and
    dashboard experiences belong to client or console repositories such as
    `rtk_cloud_admin`; those clients consume this OpenAPI contract instead of
    being implemented in this server.
...
servers:
  - url: http://localhost:8080
paths:
...
  /v1/developer/brand-clouds/{brandCloudId}/pki/test-app-certificates:
    post:
      tags: [Developer Brand Clouds]
      operationId: issueDeveloperPKITestAppCertificate
      summary: Issue a fixed 30-day certificate-only test bundle for a scoped app identity.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BrandCloudId"
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [target_type, target_id, csr_pem]
              properties:
                target_type: { type: string, enum: [user, end_user] }
                target_id: { type: string }
                csr_pem: { type: string }
      responses:
        "201":
          description: Certificate issued. The private key remains caller-managed.
          headers:
            Cache-Control:
              schema: { type: string, example: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppCertificateBootstrap" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BrandCloudId:
      name: brandCloudId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_claim_token:
              summary: Non-retryable Claim Token error.
              value:
                error:
                  code: invalid_claim_token
                  message: Invalid claim token
                  retryable: false
                  resolution_action: scan_or_enter_a_valid_claim_token
            claim_service_unavailable:
              summary: Retryable Claim Token service error.
              value:
                error:
                  code: service_unavailable
                  message: Claim resolve is temporarily unavailable
                  retryable: true
                  resolution_action: retry_later
            oidc_user_not_provisioned:
              summary: Unknown SSO user.
              value:
                error:
                  code: user_not_provisioned
                  message: SSO user is not provisioned in account manager
            oidc_disabled:
              summary: OIDC login disabled.
              value:
                error:
                  code: oidc_disabled
                  message: OIDC authentication is disabled
            invalid_oidc_state:
              summary: Invalid or replayed OIDC state.
              value:
                error:
                  code: invalid_oidc_state
                  message: Invalid OIDC login state
            invalid_oidc_token:
              summary: Invalid OIDC token.
              value:
                error:
                  code: invalid_oidc_token
                  message: Invalid OIDC token
            unverified_oidc_email:
              summary: Unverified OIDC email.
              value:
                error:
                  code: unverified_oidc_email
                  message: OIDC email is not verified
            oidc_provider_misconfigured:
              summary: OIDC provider misconfigured.
              value:
                error:
                  code: oidc_provider_misconfigured
                  message: OIDC provider is misconfigured
  schemas:
...
    AppCertificateBootstrap:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [csr_required, issued]
        subject:
          type: string
          description: APP certificate subject. Global users use `app-user:<user_id>` and APP end users use `app-end-user:<end_user_id>`.
          example: app-end-user:0f0475fb-4c8d-4f78-8b6f-3b9e76c5b34b
        certificate_pem:
          type: string
        certificate_chain_pem:
          type: string
        fingerprint_sha256:
          type: string
        serial_number:
          type: string
        issuer_request_id:
          type: string
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
        certificate_bundle:
          $ref: "#/components/schemas/RTKCertificateBundleV1"
    RTKCertificateBundleV1:
      type: object
      additionalProperties: false
      required: [format, version, profile, environment, usage, tenant_id, identity, key, certificate, issuance]
      properties:
        format: { type: string, enum: [rtk_certificate_bundle] }
        version: { type: integer, enum: [1] }
        profile: { type: string, enum: [certificate_only, test_exportable] }
        environment: { type: string, enum: [local, staging, production] }
        usage: { type: string, enum: [device_mtls, app_mtls] }
        tenant_id: { type: string }
        identity:
          type: object
          required: [kind, id, subject_dn, uri_sans]
          properties:
            kind: { type: string, enum: [device, app] }
            id: { type: string }
            subject_dn: { type: string }
            uri_sans: { type: array, items: { type: string, format: uri } }
        key:
          type: object
          required: [algorithm, spki_sha256, material]
          properties:
            algorithm: { type: string, enum: [ecdsa_p256, ed25519, rsa_2048] }
            spki_sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
            material:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [caller_managed, key_reference, embedded_pkcs8_pem] }
                provider: { type: string }
                reference: { type: string }
                private_key_pem: { type: string }
        certificate:
          type: object
          required: [chain_pem, serial_number_hex, fingerprint_sha256, not_before, not_after]
          properties:
            chain_pem: { type: array, minItems: 1, items: { type: string } }
            serial_number_hex: { type: string, pattern: "^(?:[0-9a-f]{2})+$" }
            fingerprint_sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
            not_before: { type: string, format: date-time }
            not_after: { type: string, format: date-time }
        issuance:
          type: object
          required: [request_id, issued_at]
          properties:
            request_id: { type: string }
            issued_at: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
              description: Optional machine-readable hint for Claim Token resolve failures.
            resolution_action:
              type: string
              description: Optional client action hint for Claim Token resolve failures.